+++ tests/services/file.service.js
@@ -1,0 +1,5 @@
+import test from 'ava'
+
+test.todo('FileService#scripts should return name and path of valid script')
+test.todo('FileService#modules should return name and path of valid modules')
+test.todo('FileService#modules should return target and source of each file in modules')
